refactor(ticks): split npc_respawn/ticks along tick families + test seam
Splits two oversized files under cell/service/ticks/ for #529:

- npc_respawn.rs (1097) -> npc_respawn/mod.rs (production tick, 389) +
  npc_respawn/tests.rs (the inline #[cfg(test)] mod, 707). File->dir keeps
  the same module path, so the existing super::super::super:: imports and
  the parent `mod npc_respawn;` / re-export stay valid unchanged.
- ticks/mod.rs (863) -> slim wiring/re-export mod.rs (28) plus NEW sibling
  submodules by tick family:
    - aoi.rs                run_aoi_tick
    - reload_completion.rs  reload_completion_tick
    - pending_holster.rs    pending_attack_tick + pending_reload_tick
    - regen.rs              regen_tick
  Inline #[cfg(test)] tests moved to the submodule each exercises.

Each moved sibling is one level deeper than the old mod.rs, so the
inline super::super:: references were rewritten to depth-independent
crate::cell::... absolute paths; messages/space_manager imports use the
correct super::super::super:: depth for the new location.

Behavior-preserving pure code movement. Re-export discipline keeps the
`super::ticks::<tick>` paths in message_loop unchanged. Zero
public-surface change.

Part of #529
